fix: infinite scroll + use IntersectionObserver
